restor code in base
description:restor code in base

issue:https://gitee.com/openharmony/arkcompiler_toolchain/issues/I6QN33

+++ tooling/base/pt_base64.cpp
@@ -105,7 +105,7 @@ size_t PtBase64::Encode(char *output, const void *input, std::size_t len)
         src += UNENCODED_GROUP_BYTES;
     }
 
-    int8_t remainder = len % UNENCODED_GROUP_BYTES;
+    int8_t remainder = static_cast<int8_t>(len % UNENCODED_GROUP_BYTES);
     int8_t paddingCnt = (UNENCODED_GROUP_BYTES - remainder) % UNENCODED_GROUP_BYTES;
     switch (paddingCnt) {
         case 1:
